requirements.txt
pandas
numpy
python-docx
reportlab